Extract agent config modal hooks to fix cursor-review
Move useAddConfigModalState and useEditConfigModalState into co-located
.ts files so the react-custom-hooks heuristic no longer flags multi-line
hook signatures inside TSX.
