+++ apps/hermes/dashboard/app/dashboard/agent-configs/edit-config-modal.tsx
@@ -1,8 +1,5 @@
 "use client";
 
-import { useEffect, useMemo, useState } from "react";
-import { useRouter } from "next/navigation";
-
 import {
   Dialog,
   DialogContent,
@@ -10,12 +7,11 @@ import {
   DialogTitle,
 } from "@workspace/ui/components/dialog";
 import { Button } from "@workspace/ui/components/button";
-import { useFormAction } from "@/app/dashboard/agent-configs/actions/update/.generated/use-form-action";
 
 import { AgentConfigFormFields } from "./agent-config-form-fields";
 import type { AgentConfigRow } from "./agent-config-row-actions";
 import type { VariableExpansionStringFieldLoaders } from "@workspace/variable-expansion-picker";
-import { useCloseOnSuccessfulSubmit } from "@/app/dashboard/hooks/use-close-on-successful-submit";
+import { useEditConfigModalState } from "./use-edit-config-modal-state";
 
 type AgentForDropdown = {
   id: string;
@@ -29,58 +25,6 @@ type EditConfigModalProps = {
   pickerLoaders: VariableExpansionStringFieldLoaders;
   open: boolean;
   onOpenChange: (open: boolean) => void;
-};
-
-const initialFormState = {
-  name: "",
-  description: "",
-  agentKey: "",
-  config: {} as Record<string, unknown>,
-};
-
-/**
- * Encapsulates form state synced from config, form action, and close-on-success for edit config modal.
- */
-const useEditConfigModalState = (
-  config: AgentConfigRow | null,
-  open: boolean,
-  onOpenChange: (open: boolean) => void,
-) => {
-  const router = useRouter();
-  const [formState, setFormState] = useState(initialFormState);
-  const { FormWithAction, state, pending } = useFormAction();
-
-  const errorMessage = useMemo(() => {
-    if (state && state.status === false) return state.message as string;
-    return null;
-  }, [state]);
-
-  useEffect(() => {
-    if (config && open) {
-      setFormState({
-        name: config.name,
-        description: config.description ?? "",
-        agentKey: `${config.agentId}@${config.agentVersion}`,
-        config:
-          typeof config.config === "object" && config.config !== null
-            ? { ...(config.config as Record<string, unknown>) }
-            : {},
-      });
-    }
-  }, [config, open]);
-
-  useCloseOnSuccessfulSubmit({
-    open,
-    pending,
-    state,
-    isSuccess: (nextState) => Boolean(nextState && nextState.status === true),
-    onSuccess: () => {
-      onOpenChange(false);
-      router.refresh();
-    },
-  });
-
-  return { formState, setFormState, FormWithAction, pending, errorMessage };
 };
 
 /**
